Cutting data by group

Cutting data by group

Usage

cut_by_group(df, col_data, col_group, group, cat_col)

Arguments

df

(dataframe) with a column of data to be cut and a column specifying the group of each observation.

col_data

(character) the column containing the data to be cut.

col_group

(character) the column containing the names of the groups according to which the data should be split.

group

(nested list) providing for each parameter value that should be analyzed in a categorical way: the name of the parameter (character), a series of breakpoints (numeric) where the first breakpoints is typically -Inf and the last Inf, and a series of name which will describe each category (character).

cat_col

(character) the name of the new column in which the cut label should he stored.

Value

data.frame with a column containing categorical values.

Details

Function used to categorize numeric data stored in long format depending on their group. Intervals are closed on the right (and open on the left).

Examples

group <- list(
  list(
    "Height",
    c(-Inf, 150, 170, Inf),
    c("=<150", "150-170", ">170")
  ),
  list(
    "Weight",
    c(-Inf, 65, Inf),
    c("=<65", ">65")
  ),
  list(
    "Age",
    c(-Inf, 31, Inf),
    c("=<31", ">31")
  ),
  list(
    "PreCondition",
    c(-Inf, 1, Inf),
    c("=<1", "<1")
  )
)
data <- data.frame(
  SUBJECT = rep(letters[1:10], 4),
  PARAM = rep(c("Height", "Weight", "Age", "other"), each = 10),
  AVAL = c(rnorm(10, 165, 15), rnorm(10, 65, 5), runif(10, 18, 65), rnorm(10, 0, 1)),
  index = 1:40
)

cut_by_group(data, "AVAL", "PARAM", group, "my_new_categories")
#>    SUBJECT  PARAM         AVAL index my_new_categories
#> 1        a Height 180.26701494     1              >170
#> 2        b Height 161.92811090     2           150-170
#> 3        c Height 165.02580337     3           150-170
#> 4        d Height 166.05594547     4           150-170
#> 5        e Height 177.97404941     5              >170
#> 6        f Height 149.80432443     6             =<150
#> 7        g Height 190.90612390     7              >170
#> 8        h Height 185.10230144     8              >170
#> 9        i Height 144.98886772     9             =<150
#> 10       j Height 140.68058003    10             =<150
#> 11       a Weight  66.21056693    11               >65
#> 12       b Weight  68.56469948    12               >65
#> 13       c Weight  60.94427368    13              =<65
#> 14       d Weight  59.09581103    14              =<65
#> 15       e Weight  58.34081402    15              =<65
#> 16       f Weight  60.29115615    16              =<65
#> 17       g Weight  69.27405475    17               >65
#> 18       h Weight  62.91898794    18              =<65
#> 19       i Weight  62.84010293    19              =<65
#> 20       j Weight  64.18760262    20              =<65
#> 21       a    Age  25.43451162    21              =<31
#> 22       b    Age  30.50776630    22              =<31
#> 23       c    Age  24.69835604    23              =<31
#> 24       d    Age  58.14295490    24               >31
#> 25       e    Age  41.24404000    25               >31
#> 26       f    Age  57.14280650    26               >31
#> 27       g    Age  58.33970064    27               >31
#> 28       h    Age  31.68384019    28               >31
#> 29       i    Age  24.62171889    29              =<31
#> 30       j    Age  25.38844486    30              =<31
#> 31       a  other   0.56578977    31              <NA>
#> 32       b  other   0.63897939    32              <NA>
#> 33       c  other  -0.09137375    33              <NA>
#> 34       d  other   1.07315480    34              <NA>
#> 35       e  other  -0.07183937    35              <NA>
#> 36       f  other   0.48827478    36              <NA>
#> 37       g  other  -0.26052684    37              <NA>
#> 38       h  other  -0.47987484    38              <NA>
#> 39       i  other   2.39100795    39              <NA>
#> 40       j  other  -0.35447994    40              <NA>