Cutting data by group

Cutting data by group

cut_by_group(df, col_data, col_group, group, cat_col)

Arguments

df

(dataframe) with a column of data to be cut and a column specifying the group of each observation.

col_data

(character) the column containing the data to be cut.

col_group

(character) the column containing the names of the groups according to which the data should be split.

group

(nested list) providing for each parameter value that should be analyzed in a categorical way: the name of the parameter (character), a series of breakpoints (numeric) where the first breakpoints is typically -Inf and the last Inf, and a series of name which will describe each category (character).

cat_col

(character) the name of the new column in which the cut label should he stored.

Value

data.frame with a column containing categorical values.

Details

Function used to categorize numeric data stored in long format depending on their group. Intervals are closed on the right (and open on the left).

Examples

group <- list(
  list(
    "Height",
    c(-Inf, 150, 170, Inf),
    c("=<150", "150-170", ">170")
  ),
  list(
    "Weight",
    c(-Inf, 65, Inf),
    c("=<65", ">65")
  ),
  list(
    "Age",
    c(-Inf, 31, Inf),
    c("=<31", ">31")
  ),
  list(
    "PreCondition",
    c(-Inf, 1, Inf),
    c("=<1", "<1")
  )
)
data <- data.frame(
  SUBJECT = rep(letters[1:10], 4),
  PARAM = rep(c("Height", "Weight", "Age", "other"), each = 10),
  AVAL = c(rnorm(10, 165, 15), rnorm(10, 65, 5), runif(10, 18, 65), rnorm(10, 0, 1)),
  index = 1:40
)

cut_by_group(data, "AVAL", "PARAM", group, "my_new_categories")
#>    SUBJECT  PARAM         AVAL index my_new_categories
#> 1        a Height 151.33798784     1           150-170
#> 2        b Height 167.63669711     2           150-170
#> 3        c Height 148.35467935     3             =<150
#> 4        d Height 170.57297231     4              >170
#> 5        e Height 147.53227433     5             =<150
#> 6        f Height 169.07997998     6           150-170
#> 7        g Height 169.70620394     7           150-170
#> 8        h Height 157.23110960     8           150-170
#> 9        i Height 164.14007992     9           150-170
#> 10       j Height 170.20861559    10              >170
#> 11       a Weight  66.53219238    11               >65
#> 12       b Weight  60.84837616    12              =<65
#> 13       c Weight  60.67517096    13              =<65
#> 14       d Weight  68.20757441    14               >65
#> 15       e Weight  63.35218265    15              =<65
#> 16       f Weight  57.76159210    16              =<65
#> 17       g Weight  75.30057512    17               >65
#> 18       h Weight  71.88809464    18               >65
#> 19       i Weight  60.21658240    19              =<65
#> 20       j Weight  67.88399471    20               >65
#> 21       a    Age  34.74714207    21               >31
#> 22       b    Age  39.60688353    22               >31
#> 23       c    Age  33.86506538    23               >31
#> 24       d    Age  26.68853158    24              =<31
#> 25       e    Age  56.34213383    25               >31
#> 26       f    Age  23.67289477    26              =<31
#> 27       g    Age  30.07465081    27              =<31
#> 28       h    Age  37.23764250    28               >31
#> 29       i    Age  33.72651980    29               >31
#> 30       j    Age  41.81616212    30               >31
#> 31       a  other  -0.73846315    31              <NA>
#> 32       b  other  -0.79705664    32              <NA>
#> 33       c  other   1.71064199    33              <NA>
#> 34       d  other   0.81647208    34              <NA>
#> 35       e  other  -1.07339763    35              <NA>
#> 36       f  other   0.80699302    36              <NA>
#> 37       g  other   0.61903808    37              <NA>
#> 38       h  other  -1.61546760    38              <NA>
#> 39       i  other   0.68292444    39              <NA>
#> 40       j  other   0.01529509    40              <NA>