Cutting data by group

Cutting data by group

Usage

cut_by_group(df, col_data, col_group, group, cat_col)

Arguments

df

(dataframe) with a column of data to be cut and a column specifying the group of each observation.

col_data

(character) the column containing the data to be cut.

col_group

(character) the column containing the names of the groups according to which the data should be split.

group

(nested list) providing for each parameter value that should be analyzed in a categorical way: the name of the parameter (character), a series of breakpoints (numeric) where the first breakpoints is typically -Inf and the last Inf, and a series of name which will describe each category (character).

cat_col

(character) the name of the new column in which the cut label should he stored.

Value

data.frame with a column containing categorical values.

Details

Function used to categorize numeric data stored in long format depending on their group. Intervals are closed on the right (and open on the left).

Examples

group <- list(
  list(
    "Height",
    c(-Inf, 150, 170, Inf),
    c("=<150", "150-170", ">170")
  ),
  list(
    "Weight",
    c(-Inf, 65, Inf),
    c("=<65", ">65")
  ),
  list(
    "Age",
    c(-Inf, 31, Inf),
    c("=<31", ">31")
  ),
  list(
    "PreCondition",
    c(-Inf, 1, Inf),
    c("=<1", "<1")
  )
)
data <- data.frame(
  SUBJECT = rep(letters[1:10], 4),
  PARAM = rep(c("Height", "Weight", "Age", "other"), each = 10),
  AVAL = c(rnorm(10, 165, 15), rnorm(10, 65, 5), runif(10, 18, 65), rnorm(10, 0, 1)),
  index = 1:40
)

cut_by_group(data, "AVAL", "PARAM", group, "my_new_categories")
#>    SUBJECT  PARAM        AVAL index my_new_categories
#> 1        a Height 153.1629704     1           150-170
#> 2        b Height 129.3341270     2             =<150
#> 3        c Height 178.0614270     3              >170
#> 4        d Height 181.4607255     4              >170
#> 5        e Height 158.0541546     5           150-170
#> 6        f Height 149.4678054     6             =<150
#> 7        g Height 201.2372294     7              >170
#> 8        h Height 166.4499063     8           150-170
#> 9        i Height 164.8351262     9           150-170
#> 10       j Height 164.2075257    10           150-170
#> 11       a Weight  70.8890740    11               >65
#> 12       b Weight  63.3183055    12              =<65
#> 13       c Weight  69.3625435    13               >65
#> 14       d Weight  58.3580305    14              =<65
#> 15       e Weight  73.2355160    15               >65
#> 16       f Weight  62.0452534    16              =<65
#> 17       g Weight  70.8721084    17               >65
#> 18       h Weight  61.7141056    18              =<65
#> 19       i Weight  69.3952961    19               >65
#> 20       j Weight  60.6072826    20              =<65
#> 21       a    Age  58.1326049    21               >31
#> 22       b    Age  52.4674769    22               >31
#> 23       c    Age  43.7397763    23               >31
#> 24       d    Age  32.1927012    24               >31
#> 25       e    Age  41.3387877    25               >31
#> 26       f    Age  54.2420382    26               >31
#> 27       g    Age  27.3003088    27              =<31
#> 28       h    Age  21.5727597    28              =<31
#> 29       i    Age  62.6027660    29               >31
#> 30       j    Age  31.3133930    30               >31
#> 31       a  other  -0.8632284    31              <NA>
#> 32       b  other  -0.9601284    32              <NA>
#> 33       c  other   1.8060664    33              <NA>
#> 34       d  other   0.8748184    34              <NA>
#> 35       e  other  -0.8400238    35              <NA>
#> 36       f  other  -0.3800949    36              <NA>
#> 37       g  other  -2.1026478    37              <NA>
#> 38       h  other  -1.4036098    38              <NA>
#> 39       i  other   0.8480489    39              <NA>
#> 40       j  other   0.3721646    40              <NA>